home-page layout and routing
