--- src/frontend/App.jsx
@@ -1,11 +1,0 @@
-import { useState } from "react";
-import Home from './pages/home';
-
-function App() {
-
-  return (
-    <Home />
-  );
-}
-
-export default App;
