+++ src/frontend/routes/Browse.jsx
@@ -1,0 +1,5 @@
+const Browse = () => {
+
+};
+
+export default Browse;
