+++ src/frontend/routes/Profile.jsx
@@ -1,0 +1,5 @@
+const Profile = () => {
+
+};
+
+export default Profile;
