+++ src/frontend/styles/global.css
@@ -7,3 +7,15 @@
 html {
   background-color: #202430;
 }
+
+svg path {
+  stroke: currentColor;
+}
+
+svg rect {
+  stroke: currentColor;
+}
+
+svg circle {
+  stroke: currentColor;
+}
